requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
kafka-python==1.4.2
redis==3.0.1
requests>=2.20.0
schedule==0.6.0
urllib3>=1.24.2
pytz==2018.9
python-dateutil==2.8.0
pylint==2.4.2
sphinx==2.2.0
sphinx-rtd-theme==0.4.3